스킬마다 사용하는 mp변경, 그리고 초당 hp 5, mp는 10 회복

+++ Witle/Status.cpp
@@ -5,13 +5,13 @@
 #include "GameObject.h"
 #include "Status.h"
   
-void Status::SubstractHP(UINT damage)
+void Status::SubstractHP(float damage)
 { 
 	m_Guage -= damage;
 	if (m_Guage < 0.f) { m_Guage = 0.f; };
 }
 
-void Status::AddHP(UINT damage)
+void Status::AddHP(float damage)
 {
 	m_Guage += damage;
 	if (m_Guage > m_MAXGuage) { m_Guage = m_MAXGuage; };
